aadikarta-social-media/image_service/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
aiohttp
tenacity
minio
python-multipart
structlog
